Cutting Mechanic Bug Fix
I check if start mouse position and end mouse position is same, if it is I was not allowing players to cut the fruit but also I was assigning the fruit position to the start pos after cutting it to prevent some problems. Current problem was start pos and end pos wouldn't be the same  even I assigned the start position to make sure players move the mouse or finger. I fixed this issue by assigning end pos to start pos.